(BE) added logic for index cards

+++ knowledge-base-be/src/main/java/com/ondra/knowledgebasebe/topic/TopicService.java
@@ -4,6 +4,7 @@ import com.ondra.knowledgebasebe.doc.DocRepository;
 import com.ondra.knowledgebasebe.exceptions.ConstraintViolationException;
 import com.ondra.knowledgebasebe.exceptions.TopicNameAlreadyTakenException;
 import com.ondra.knowledgebasebe.exceptions.TopicNotFoundException;
+import com.ondra.knowledgebasebe.indexcard.IndexCardRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,10 +15,12 @@ public class TopicService {
 
     private final TopicRepository topicRepository;
     private final DocRepository docRepository;
+    private final IndexCardRepository indexCardRepository;
 
-    public TopicService(TopicRepository topicRepository, DocRepository docRepository) {
+    public TopicService(TopicRepository topicRepository, DocRepository docRepository, IndexCardRepository indexCardRepository) {
         this.topicRepository = topicRepository;
         this.docRepository = docRepository;
+        this.indexCardRepository = indexCardRepository;
     }
 
     public TopicDto addTopic(String name) {
@@ -39,6 +42,8 @@ public class TopicService {
         if (!topicRepository.existsById(id)) throw new TopicNotFoundException(id);
         if (docRepository.existsByTopicId(id))
             throw new ConstraintViolationException("Topic with ID " + id + " cannot be deleted as long as documentations exist that are linked to this topic");
+        if (indexCardRepository.existsByTopicId(id))
+            throw new ConstraintViolationException("Topic with ID " + id + " cannot be deleted as long as index cards exist that are linked to this topic");
         topicRepository.deleteById(id);
     }
